one more unit test refactor
